# target/classes/openapi.yaml
openapi: 3.0.1
info:
  description: This specification is intended to describe the functionality of the
    photo album application
  title: Photo album OpenAPI specification
  version: 0.0.1
servers:
- description: Dev server
  url: http://localhost:8080
paths:
  /users/login:
    get:
      operationId: loginUser
      parameters:
      - description: The email for login
        explode: true
        in: query
        name: email
        required: true
        schema:
          type: string
        style: form
      - description: The password for login
        explode: true
        in: query
        name: password
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Method for user authorization
      tags:
      - AuthController
      x-accepts: application/json
      x-tags:
      - tag: AuthController
  /users/registr:
    post:
      operationId: registrUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDto'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReadDto'
          description: Successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Method for registration
      tags:
      - AuthController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: AuthController
  /users/refresh:
    post:
      description: refresh user's auth
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshDto'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Method for refresh token
      tags:
      - AuthController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: AuthController
  /users/forgot-password:
    put:
      description: after this method user should check his email
      operationId: processForgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        default:
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Forgot password
      tags:
      - AuthController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: AuthController
  /users/reset-password:
    put:
      description: Method for resete password
      operationId: processResetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        default:
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: something wrong
      summary: Reset password
      tags:
      - AuthController
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: AuthController
...
components:
  schemas:
    UserCreateDto:
      example:
        password: MyNew_pass
        name: theUser
        email: egor@email.com
      properties:
        name:
          example: theUser
          type: string
        email:
          example: egor@email.com
          type: string
        password:
          example: MyNew_pass
          type: string
      type: object
    UserReadDto:
      example:
        password: MyNew_pass
        name: theUser
        id: 0
        email: egor@email.com
      properties:
        id:
          format: int64
          type: integer
        name:
          example: theUser
          type: string
        email:
          example: egor@email.com
          type: string
        password:
          example: MyNew_pass
          type: string
      type: object
    ForgotPasswordDto:
      example:
        email: egor@email.com
      properties:
        email:
          example: egor@email.com
          type: string
      type: object
    UpdatePasswordDto:
      example:
        password: password
        token: token
      properties:
        token:
          type: string
        password:
          type: string
      type: object
    JwtResponse:
      example:
        id: 0
        accessToken: accessToken
        email: egor@email.com
        refreshToken: refreshToken
      properties:
        id:
          format: int64
          type: integer
        email:
          example: egor@email.com
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
      type: object
    RefreshDto:
      example:
        refreshToken: refreshToken
      properties:
        refreshToken:
          type: string
      type: object
...
    Error:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object